Correct namespacing for a SimpleCov formatter
The convention for SimpleCov formatters is to be in SimpleCov::Formatter, but I
created it in SimpleCov::Inline.

Example:
https://github.com/simplecov-ruby/simplecov-html/blob/7c8d0bc9fbdb60f8f8ad00e6b79f2403cf2dda98/lib/simplecov-html.rb

Move files around to make it right.
