+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## [0.1.0] - 2024-07-19
 
 - Initial release
+
+## [1.0.0] - 2024-07-19
+
+- Fix namespacing
