+++ README.md
@@ -1,4 +1,4 @@
-# SimpleCov::Inline
+# SimpleCov::Formatter::Inline
 
 Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/simplecov/inline`. To experiment with that code, run `bin/console` for an interactive prompt.
 
@@ -34,4 +34,4 @@ The gem is available as open source under the terms of the [MIT License](https:/
 
 ## Code of Conduct
 
-Everyone interacting in the SimpleCov::Inline project's codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/appbot/simplecov-inline/blob/main/CODE_OF_CONDUCT.md).
+Everyone interacting in the SimpleCov::Formatter::Inline project's codebases, issue trackers, chat rooms and mailing lists is expected to follow the [code of conduct](https://github.com/appbot/simplecov-inline/blob/main/CODE_OF_CONDUCT.md).
