+++ lib/simplecov/formatter/inline/version.rb
@@ -1,0 +1,7 @@
+module SimpleCov
+  module Formatter
+    class Inline
+      VERSION = '1.0.0'.freeze
+    end
+  end
+end
