+++ lib/simplecov/inline.rb
@@ -1,11 +1,4 @@
-require_relative 'inline/version'
-require_relative 'inline/formatter'
-require_relative 'inline/rspec_formatter_skip_on_failure'
-require_relative 'inline/integration'
-
-module SimpleCov
-  module Inline
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
+require_relative 'formatter/inline'
+require_relative 'formatter/inline/integration'
+require_relative 'formatter/inline/rspec_formatter_skip_on_failure'
+require_relative 'formatter/inline/version'
