--- lib/simplecov/inline/version.rb
@@ -1,5 +1,0 @@
-module SimpleCov
-  module Inline
-    VERSION = '0.1.0'.freeze
-  end
-end
