+++ spec/lib/simplecov/formatter/inline/integration_spec.rb
@@ -1,5 +1,5 @@
-RSpec.describe SimpleCov::Inline::Integration do
-  after { SimpleCov::Inline::Formatter.reset_config }
+RSpec.describe SimpleCov::Formatter::Inline::Integration do
+  after { SimpleCov::Formatter::Inline.reset_config }
 
   describe '#configure_rspec_rails' do
     subject { described_class.configure_rspec_rails(rspec:, rails:) }
@@ -22,13 +22,15 @@ RSpec.describe SimpleCov::Inline::Integration do
 
       expect(RSpec::Core::Formatters)
         .to have_received(:register)
-        .with(SimpleCov::Inline::RSpecFormatterSkipOnFailure, :dump_failures)
+        .with(SimpleCov::Formatter::Inline::RSpecFormatterSkipOnFailure, :dump_failures)
     end
 
     it 'adds a formatter to the rspec config', :aggregate_failures do
       subject
 
-      expect(rspec_config).to have_received(:add_formatter).with(SimpleCov::Inline::RSpecFormatterSkipOnFailure)
+      expect(rspec_config)
+        .to have_received(:add_formatter)
+        .with(SimpleCov::Formatter::Inline::RSpecFormatterSkipOnFailure)
     end
 
     it 'adds a before suite hook that calls configure_formatter', :aggregate_failures do
@@ -57,11 +59,11 @@ RSpec.describe SimpleCov::Inline::Integration do
       let(:pattern) { '**{,/*/**}/*_spec.rb' }
 
       it 'does not filter files' do
-        expect { subject }.not_to change { SimpleCov::Inline::Formatter.config.files }.from(nil)
+        expect { subject }.not_to change { SimpleCov::Formatter::Inline.config.files }.from(nil)
       end
 
       it 'does not supress output' do
-        expect { subject }.not_to change { SimpleCov::Inline::Formatter.config.no_output }.from(nil)
+        expect { subject }.not_to change { SimpleCov::Formatter::Inline.config.no_output }.from(nil)
       end
     end
 
@@ -69,12 +71,12 @@ RSpec.describe SimpleCov::Inline::Integration do
       let(:files_to_run) { ["#{rails_root}/spec/lib/lib_spec.rb"] }
 
       it 'does not supress output' do
-        expect { subject }.not_to change { SimpleCov::Inline::Formatter.config.no_output }.from(nil)
+        expect { subject }.not_to change { SimpleCov::Formatter::Inline.config.no_output }.from(nil)
       end
 
       it 'filters by the spec file and the file it is testing' do
         expect { subject }
-          .to change { SimpleCov::Inline::Formatter.config.files }
+          .to change { SimpleCov::Formatter::Inline.config.files }
           .from(nil).to(["#{rails_root}/spec/lib/lib_spec.rb", "#{rails_root}/lib/lib.rb"])
       end
     end
@@ -83,12 +85,12 @@ RSpec.describe SimpleCov::Inline::Integration do
       let(:files_to_run) { ["#{rails_root}/spec/models/model_spec.rb"] }
 
       it 'does not supress output' do
-        expect { subject }.not_to change { SimpleCov::Inline::Formatter.config.no_output }.from(nil)
+        expect { subject }.not_to change { SimpleCov::Formatter::Inline.config.no_output }.from(nil)
       end
 
       it 'filters by the spec file and the file it is testing' do
         expect { subject }
-          .to change { SimpleCov::Inline::Formatter.config.files }
+          .to change { SimpleCov::Formatter::Inline.config.files }
           .from(nil).to(["#{rails_root}/spec/models/model_spec.rb", "#{rails_root}/app/models/model.rb"])
       end
     end
@@ -104,13 +106,13 @@ RSpec.describe SimpleCov::Inline::Integration do
 
       it 'supresses output' do
         expect { subject }
-          .to change { SimpleCov::Inline::Formatter.config.no_output }
+          .to change { SimpleCov::Formatter::Inline.config.no_output }
           .from(nil)
           .to('filtered to line of code')
       end
 
       it 'does not filter files' do
-        expect { subject }.not_to change { SimpleCov::Inline::Formatter.config.files }.from(nil)
+        expect { subject }.not_to change { SimpleCov::Formatter::Inline.config.files }.from(nil)
       end
     end
 
@@ -126,13 +128,13 @@ RSpec.describe SimpleCov::Inline::Integration do
       let(:files_to_run) { ["#{rails_root}/spec/models"] }
 
       it 'does not supress output' do
-        expect { subject }.not_to change { SimpleCov::Inline::Formatter.config.no_output }.from(nil)
+        expect { subject }.not_to change { SimpleCov::Formatter::Inline.config.no_output }.from(nil)
       end
 
       it 'only filters to the spec directory and does inlcude the covered files' do
         # Existing behaviour. Feels like it should suppress instead.
         expect { subject }
-          .to change { SimpleCov::Inline::Formatter.config.files }
+          .to change { SimpleCov::Formatter::Inline.config.files }
           .from(nil).to(["#{rails_root}/spec/models"])
       end
     end
