--- spec/simplecov/inline_spec.rb
@@ -1,5 +1,0 @@
-RSpec.describe SimpleCov::Inline do
-  it 'has a version number' do
-    expect(SimpleCov::Inline::VERSION).not_to be_nil
-  end
-end
